Refactor main.tsx and add Redux store
Refactor main.tsx to include the Redux store and Provider component for state management. This allows for better organization and management of application state. The Redux store is configured in store.ts, which is a new file added to the project. This change sets up the foundation for future state management using Redux.

Add drawerSlice.ts for drawer state management

Add drawerSlice.ts, a new file that defines the drawerSlice reducer for managing the state of the drawer component. This reducer handles the toggleDrawer action, which toggles the open state of the drawer. This addition enables better control and management of the drawer component's state.